move version to a single place (setup.cfg) and read from metadata for in-lib usage
