+++ restpite/__init__.py
@@ -1,13 +1,25 @@
 """isort:skip_file"""
 
+import importlib.metadata
+
+__version__ = importlib.metadata.version("restpite")
+
+
 from .http.response import RestpiteResponse
 from .protocols.restpite_protocols import Notifyable
 from .http.session import RestpiteSession
 from .http.request import RestpiteRequest
+
 
 import logging
 
 log = logging.getLogger(__name__)
 log.addHandler(logging.NullHandler())
 
-__all__ = ["RestpiteSession", "RestpiteRequest", "RestpiteResponse", "Notifyable"]
+__all__ = [
+    "RestpiteSession",
+    "RestpiteRequest",
+    "RestpiteResponse",
+    "Notifyable",
+    "__version__",
+]
