--- restpite/__version__.py
@@ -1,16 +1,0 @@
-#  ____           _         _ _
-# |  _ \ ___  ___| |_ _ __ (_) |_ ___
-# | |_) / _ \/ __| __| '_ \| | __/ _ \
-# |  _ <  __/\__ \ |_| |_) | | ||  __/
-# |_| \_\___||___/\__| .__/|_|\__\___|
-#                    |_|
-
-__title__ = "restpite"
-__description__ = "Python DSL for testing restful APIs"
-__url__ = "https://restpite.readthedocs.io/en/latest/readme.html"
-__version__ = "0.1.1"
-__build__ = None
-__author__ = "Simon Kerr"
-__author_email__ = "jackofspaces@gmail.com"
-__license__ = "MIT"
-__copyright__ = "Copyright 2020 Simon Kerr"
