+++ setup.cfg
@@ -1,6 +1,13 @@
+#  ____           _         _ _
+# |  _ \ ___  ___| |_ _ __ (_) |_ ___
+# | |_) / _ \/ __| __| '_ \| | __/ _ \
+# |  _ <  __/\__ \ |_| |_) | | ||  __/
+# |_| \_\___||___/\__| .__/|_|\__\___|
+#                    |_|
+
 [metadata]
 name = restpite
-version = 0.1.1
+version = 1.1.2
 description = Python DSL for easily testing REST services
 author = Simon Kerr
 author_email = jackofspaces@gmail.com
