+++ tox.ini
@@ -2,7 +2,7 @@
 envlist =
             clean,
             py38,
-            codecov
+
 ignore_basepython_conflict = true
 
 [testenv]
